# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sistema de carteiras digitais multi-tenant desenvolvido em Go",
        "title": "Suno Wallets API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte API",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/b3/admin/reactivation/analyze": {
            "post": {
                "description": "Analisa a situação de sincronização de um cliente e sugere a melhor estratégia para reativação, considerando gaps de dados e última sincronização.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B3 Reactivation"
                ],
                "summary": "Analisar estratégia de reativação (Admin)",
                "parameters": [
                    {
                        "type": "string",
                        "example": "status_invest",
                        "description": "ID do inquilino (tenant)",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Parâmetros de análise",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plano de reativação sugerido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Parâmetros inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro na análise",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/b3/admin/reactivation/execute": {
            "post": {
                "description": "Executa a estratégia de reativação previamente analisada, sincronizando dados faltantes do cliente com a B3. Suporta modo dry-run para simulação.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B3 Reactivation"
                ],
                "summary": "Executar plano de reativação (Admin)",
                "parameters": [
                    {
                        "type": "string",
                        "example": "status_invest",
                        "description": "ID do inquilino (tenant)",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Parâmetros de execução",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado da execução",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Parâmetros inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro na execução",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/b3/client/reactivation/status": {
            "get": {
                "description": "Verifica o status atual de sincronização do cliente e informa se há necessidade de reativação, sem executar operações.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "B3 Reactivation"
                ],
                "summary": "Status de reativação do cliente",
                "parameters": [
                    {
                        "type": "string",
                        "example": "status_invest",
                        "description": "ID do inquilino (tenant)",
                        "name": "X-Tenant-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "12345678901",
                        "description": "CPF do cliente (11 dígitos, apenas números)",
                        "name": "cpf",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status de reativação",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "CPF inválido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro na verificação",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "TenantID": {
            "description": "ID do inquilino para multi-tenancy",
            "type": "apiKey",
            "name": "X-Tenant-ID",
            "in": "header"
        }
    }
}